requirements.txt
antlr4-python3-runtime==4.8
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
filelock==3.4.2
Flask==2.0.2
gensim==4.1.2
huggingface-hub==0.4.0
idna==3.3
itsdangerous==2.0.1
gunicorn==20.1.0 
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
py-sudoku==1.0.2
pyparsing==3.0.7
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.16.2
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
